FreeRTOS with Bluetooth on Tracking Car

+++ V2V-secondVehicleECU/OTHERS/std_types.h
@@ -8,8 +8,8 @@ typedef unsigned char 						u8;
 
 typedef signed short int 					s16;
 typedef unsigned short int 					u16;
-typedef signed long 						s32;
-typedef unsigned long 						u32;
+typedef signed /*long*/ int						s32;
+typedef unsigned /*long*/ int					u32;
 typedef signed long long int  				s64;
 typedef unsigned long long int  			u64;
